services/process-worker/requirements.txt
psycopg2-binary==2.9.9
yt-dlp==2025.06.25
ffmpeg